# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: STRESSNES luxury fashion ecommerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with filters and pagination
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            default: 24
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [DRAFT, ACTIVE, ARCHIVED]
        - in: query
          name: categoryId
          schema:
            type: string
        - in: query
          name: collectionId
          schema:
            type: string
        - in: query
          name: featured
          schema:
            type: boolean
        - in: query
          name: published
          schema:
            type: boolean
        - in: query
          name: minPrice
          schema:
            type: number
        - in: query
          name: maxPrice
          schema:
            type: number
        - in: query
          name: sortBy
          schema:
            type: string
            enum: [price, createdAt, title]
            default: createdAt
        - in: query
          name: sortOrder
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        "200":
          description: Paginated product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedProductList"

    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{slug}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by slug
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found

    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product (admin only)
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (admin only)
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted

  # ── Categories ──────────────────────────────────────────────
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (own orders for customer, all for admin)
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            default: 20
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: Paginated order list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedOrderList"

    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order from the current cart
      parameters:
        - in: query
          name: cartId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update an order's status (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusInput"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ── Customers ───────────────────────────────────────────────
...
components:
  schemas:
...
    ProductImage:
      type: object
      required: [id, url]
      properties:
        id:
          type: string
        url:
          type: string
        altText:
          type: ["string", "null"]
        isPrimary:
          type: boolean
        sortOrder:
          type: integer
    ProductVariant:
      type: object
      required: [id, sku, isActive]
      properties:
        id:
          type: string
        sku:
          type: string
        size:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        material:
          type: ["string", "null"]
        priceOverride:
          type: ["number", "null"]
        isActive:
          type: boolean
        stockQty:
          type: integer
    ProductCategoryRef:
      type: object
      required: [id, name, slug]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
    Product:
      type: object
      required: [id, title, slug, sku, price, status]
      properties:
        id:
          type: string
        title:
          type: string
        slug:
          type: string
        sku:
          type: string
        description:
          type: ["string", "null"]
        shortDescription:
          type: ["string", "null"]
        price:
          type: number
        comparePrice:
          type: ["number", "null"]
        status:
          type: string
        featured:
          type: boolean
        published:
          type: boolean
        images:
          type: array
          items:
            $ref: "#/components/schemas/ProductImage"
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariant"
        category:
          oneOf:
            - $ref: "#/components/schemas/ProductCategoryRef"
            - type: "null"
        collection:
          oneOf:
            - $ref: "#/components/schemas/ProductCategoryRef"
            - type: "null"
        createdAt:
          type: string
          format: date-time
    PaginatedProductList:
      type: object
      required: [data, total, page, pageSize, totalPages]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer
    ProductInput:
      type: object
      required: [title, sku, price]
      properties:
        title:
          type: string
        slug:
          type: string
        sku:
          type: string
        description:
          type: string
        shortDescription:
          type: string
        price:
          type: number
        comparePrice:
          type: number
        costPrice:
          type: number
        status:
          type: string
          enum: [DRAFT, ACTIVE, ARCHIVED]
        featured:
          type: boolean
        published:
          type: boolean
        categoryId:
          type: string
        collectionId:
          type: string
        tagIds:
          type: array
          items:
            type: string
    ProductUpdate:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        shortDescription:
          type: string
        price:
          type: number
        comparePrice:
          type: number
        status:
          type: string
          enum: [DRAFT, ACTIVE, ARCHIVED]
        featured:
          type: boolean
        published:
          type: boolean
        categoryId:
          type: string
        collectionId:
          type: string

    # ── Categories ─────────────────────────────────────────────
...
    OrderItem:
      type: object
      required: [id, productId, quantity, unitPrice, totalPrice]
      properties:
        id:
          type: string
        productId:
          type: string
        variantId:
          type: ["string", "null"]
        productTitle:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        totalPrice:
          type: number
    Order:
      type: object
      required: [id, orderNumber, status, subtotal, total, createdAt]
      properties:
        id:
          type: string
        orderNumber:
          type: string
        status:
          type: string
        subtotal:
          type: number
        discountAmount:
          type: number
        taxAmount:
          type: number
        shippingCost:
          type: number
        total:
          type: number
        currency:
          type: string
        shippingAddress:
          type: object
        couponCode:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
          format: date-time
    PaginatedOrderList:
      type: object
      required: [data, total, page, pageSize, totalPages]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer
    OrderInput:
      type: object
      required: [shippingAddress]
      properties:
        shippingAddress:
          type: object
          required: [firstName, lastName, line1, city, country]
          properties:
            firstName:
              type: string
            lastName:
              type: string
            line1:
              type: string
            line2:
              type: string
            city:
              type: string
            state:
              type: string
            postalCode:
              type: string
            country:
              type: string
            phone:
              type: string
        shippingMethodId:
          type: string
        couponCode:
          type: string
        notes:
          type: string
    OrderStatusInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [PENDING, PAID, PROCESSING, PACKED, SHIPPED, DELIVERED, CANCELLED, REFUNDED]
        note:
          type: string

    # ── Customers ──────────────────────────────────────────────
...