# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * STRESSNES luxury fashion ecommerce API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductInputStatus } from './productInputStatus';

export interface ProductInput {
  title: string;
  slug?: string;
  sku: string;
  description?: string;
  shortDescription?: string;
  price: number;
  comparePrice?: number;
  costPrice?: number;
  status?: ProductInputStatus;
  featured?: boolean;
  published?: boolean;
  categoryId?: string;
  collectionId?: string;
  tagIds?: string[];
}